add patch func, refactor code

+++ mmdet/core/loss/losses.py
@@ -8,14 +8,14 @@ from ...ops import sigmoid_focal_loss
 def weighted_nll_loss(pred, label, weight, avg_factor=None):
     if avg_factor is None:
         avg_factor = max(torch.sum(weight > 0).float().item(), 1.)
-    raw = F.nll_loss(pred.float(), label, reduction='none')
+    raw = F.nll_loss(pred, label, reduction='none')
     return torch.sum(raw * weight)[None] / avg_factor
 
 
 def weighted_cross_entropy(pred, label, weight, avg_factor=None, reduce=True):
     if avg_factor is None:
         avg_factor = max(torch.sum(weight > 0).float().item(), 1.)
-    raw = F.cross_entropy(pred.float(), label, reduction='none')
+    raw = F.cross_entropy(pred, label, reduction='none')
     if reduce:
         return torch.sum(raw * weight)[None] / avg_factor
     else:
@@ -28,7 +28,7 @@ def weighted_binary_cross_entropy(pred, label, weight, avg_factor=None):
     if avg_factor is None:
         avg_factor = max(torch.sum(weight > 0).float().item(), 1.)
     return F.binary_cross_entropy_with_logits(
-        pred.float(), label.float(), weight.float(),
+        pred, label.float(), weight.float(),
         reduction='sum')[None] / avg_factor
 
 
@@ -38,13 +38,13 @@ def py_sigmoid_focal_loss(pred,
                           gamma=2.0,
                           alpha=0.25,
                           reduction='mean'):
-    pred_sigmoid = pred.float().sigmoid()
+    pred_sigmoid = pred.sigmoid()
     target = target.type_as(pred)
     pt = (1 - pred_sigmoid) * target + pred_sigmoid * (1 - target)
     weight = (alpha * target + (1 - alpha) * (1 - target)) * weight
     weight = weight * pt.pow(gamma)
     loss = F.binary_cross_entropy_with_logits(
-        pred.float(), target, reduction='none') * weight
+        pred, target, reduction='none') * weight
     reduction_enum = F._Reduction.get_enum(reduction)
     # none: 0, mean:1, sum: 2
     if reduction_enum == 0:
@@ -65,14 +65,14 @@ def weighted_sigmoid_focal_loss(pred,
     if avg_factor is None:
         avg_factor = torch.sum(weight > 0).float().item() / num_classes + 1e-6
     return torch.sum(
-        sigmoid_focal_loss(pred.float(), target, gamma, alpha, 'none') *
+        sigmoid_focal_loss(pred, target, gamma, alpha, 'none') *
         weight.view(-1, 1))[None] / avg_factor
 
 
 def mask_cross_entropy(pred, target, label):
     num_rois = pred.size()[0]
     inds = torch.arange(0, num_rois, dtype=torch.long, device=pred.device)
-    pred_slice = pred[inds, label].squeeze(1).float()
+    pred_slice = pred[inds, label].squeeze(1)
     return F.binary_cross_entropy_with_logits(
         pred_slice, target, reduction='mean')[None]
 
@@ -80,7 +80,7 @@ def mask_cross_entropy(pred, target, label):
 def smooth_l1_loss(pred, target, beta=1.0, reduction='mean'):
     assert beta > 0
     assert pred.size() == target.size() and target.numel() > 0
-    diff = torch.abs(pred.float() - target)
+    diff = torch.abs(pred - target)
     loss = torch.where(diff < beta, 0.5 * diff * diff / beta,
                        diff - 0.5 * beta)
     reduction_enum = F._Reduction.get_enum(reduction)
